Some sketching of how the table syntax will look.
--HG--
branch : master
