+++ tinycowc/mkpat.lua
@@ -1,0 +1,7 @@
+require "./libcowgol.lua"
+
+local args = {...}
+local midcodetab = args[2]
+local archpat = args[3]
+
+local midcodes = loadmidcodes(midcodetab)
